Muevo la parte de autenticacion a un blueprint como hicimos en auth server
